gcp/requirements.txt
google-cloud-aiplatform>=1.45.0
google-cloud-storage>=2.0.0
flask>=3.0.0
gunicorn>=21.0.0
